concat controlle files
